Consolidated eventviewer code
